refactor: move sass compilation to build script (refs SFKUI-6500)
