+++ packages/theme-desktop/build.mjs
@@ -1,7 +1,11 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import { parse } from "@adobe/css-tools";
+import { moduleImporter } from "@forsakringskassan/sass-module-importer";
+import cssnano from "cssnano";
 import { getVariables } from "get-css-variables";
+import postcss from "postcss";
+import * as sass from "sass";
 import { extractSassVariables } from "./scripts/extract-sass-variables.mjs";
 
 /**
@@ -77,6 +81,52 @@ function* getTokens(rule, palette) {
 }
 
 /**
+ * @returns {Promise<void>}
+ */
+async function generateVersion(dst) {
+    const content = [
+        "// This is a generated file. See build.mjs",
+        `$version: "${process.env.npm_package_version}";`,
+        "",
+    ].join("\n");
+    await fs.writeFile(dst, content, "utf-8");
+}
+
+async function postprocess(css, from, to) {
+    const cssnanoOptions = {
+        preset: [
+            "default",
+            {
+                discardComments: { removeAll: true },
+                normalizeWhitespace: false,
+            },
+        ],
+    };
+    const plugins = [cssnano(cssnanoOptions)].filter(Boolean);
+    return await postcss(plugins).process(css, { from, to });
+}
+
+/**
+ * @param {string} src
+ * @param {string} dst
+ */
+export async function compileSass(src, dst) {
+    const data = /* SCSS */ `
+        @use "${src}"
+    `;
+    const result = sass.compileString(data, {
+        style: "expanded",
+        importers: [moduleImporter],
+        loadPaths: ["."],
+    });
+
+    const production = await postprocess(result.css, src, dst);
+    const { dir, name } = path.parse(dst);
+    const cssFile = path.join(dir, `${name}.css`);
+    await fs.writeFile(cssFile, production.css, "utf8");
+}
+
+/**
  * @param {string} pattern
  * @returns {Promise<void>}
  */
@@ -97,6 +147,29 @@ async function extractCssVariables(pattern) {
         console.log(file, "->", `dist/${name}.js`);
         await fs.writeFile(`dist/${name}.js`, cjs, "utf-8");
         await fs.writeFile(`dist/${name}.d.ts`, cts, "utf-8");
+    }
+
+    console.groupEnd();
+}
+
+/**
+ * @param {string} pattern
+ * @param {string} dir
+ * @returns {Promise<void>}
+ */
+async function buildCss(pattern, dir) {
+    console.group("Compiling CSS:");
+
+    await generateVersion("src/_version.scss");
+    for await (const item of fs.glob(pattern)) {
+        const src = item.replace(/\\/g, "/");
+        const { name } = path.parse(src);
+        if (name.startsWith("_")) {
+            continue;
+        }
+        const dst = `${dir}/${name}.css`;
+        console.log(src, "->", dst);
+        await compileSass(src, dst);
     }
 
     console.groupEnd();
@@ -126,11 +199,10 @@ async function buildMetadata(src, dst) {
     await fs.writeFile(`${dst}.d.mts`, metadataDts, "utf-8");
 }
 
+await fs.rm("dist", { recursive: true, force: true });
+await fs.rm("src/_version.scss", { force: true });
+await fs.mkdir("dist");
+await buildCss("src/*.scss", "dist");
 await extractSassVariables("src/palette-variables.scss", "dist/palette.json");
 await extractCssVariables("dist/*.css");
 await buildMetadata("./dist/theme-light.css", "./dist/metadata");
-
-await fs.copyFile(
-    "src/deprecated-variables.json",
-    "dist/deprecated-variables.json",
-);
