--- packages/theme-desktop/generate-version.js
@@ -1,19 +1,0 @@
-const fs = require("node:fs/promises");
-const path = require("node:path");
-
-/**
- * @returns {string}
- */
-function getContent() {
-    return `// This is a generated file. See generate-version.js
-$version: "${process.env.npm_package_version}";
-`;
-}
-
-async function generateVersion() {
-    const dst = path.join(__dirname, "src", "_version.scss");
-    const content = getContent();
-    await fs.writeFile(dst, content, "utf-8");
-}
-
-generateVersion();
